Load plugins into classpath in bootstrap
